+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 1.5.1 (not released yet)
 --------------------------------
 - fix alpha draw order issues in envirosuit1.vcm, Vet1.v3m, coffeesmokedtbl2.v3m, coffeesmokedtblAlt.v3m meshes
 - add support for bitmaps with alpha channel in Display_Fullscreen_Image event
+- fix editor not using all space for rendering when used with a big monitor
 
 Version 1.5.0 (released 2020-09-05)
 -----------------------------------